Fix whitespace issues.
